backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
qdrant-client==1.7.0
minio==7.2.0
whisper==20231117
edge-tts==6.1.10
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0
